requirements.txt
# Core dependencies
pandas>=2.2.2
numpy>=1.24.0
openpyxl>=3.1.0
python-dotenv>=1.0.1

# Streamlit dashboard
streamlit>=1.38.0
plotly>=5.18.0

# API and data processing
requests>=2.32.3
python-dateutil>=2.9.0
pyarrow>=17.0.0

# Statistical analysis
scipy>=1.11.0

# Database (optional, for future use)
sqlalchemy
psycopg2-binary
boto3
